Move HeaderField to common components
