+++ source/frontend/src/features/properties/map/activity/detail/ActivityHeader.tsx
@@ -1,6 +1,6 @@
+import { HeaderField } from 'components/common/HeaderField/HeaderField';
 import { StyledDivider } from 'components/common/styles';
 import { UserNameTooltip } from 'components/common/UserNameTooltip';
-import { HeaderField } from 'features/mapSideBar/tabs/HeaderField';
 import { Api_Activity } from 'models/api/Activity';
 import { Api_File } from 'models/api/File';
 import * as React from 'react';
