+++ source/frontend/src/features/properties/map/research/common/ResearchHeader.tsx
@@ -1,5 +1,5 @@
+import { HeaderField } from 'components/common/HeaderField/HeaderField';
 import { UserNameTooltip } from 'components/common/UserNameTooltip';
-import { HeaderField } from 'features/mapSideBar/tabs/HeaderField';
 import { Api_ResearchFile } from 'models/api/ResearchFile';
 import Api_TypeCode from 'models/api/TypeCode';
 import * as React from 'react';
